requirements.txt
# RAG Pix Regulation - Core dependencies
# Python 3.10+ required

# Embeddings & retrieval
sentence-transformers>=3.0.0,<4.0.0
transformers>=4.30.0
weaviate-client>=4.0.0,<5.0.0

# Observability (Phoenix + OpenInference)
arize-phoenix>=13.0.0
openinference-instrumentation>=0.1.0
openinference-instrumentation-openai>=0.1.0

# Application
streamlit>=1.28.0,<2.0.0

# LLM (Ollama for local development)
ollama>=0.3.0

# PDF parsing
pymupdf>=1.24.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Data & utilities
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
pyyaml>=6.0.0
python-json-logger>=2.0.0
tqdm>=4.65.0
